[Navigator] Fix Mac UX Issues for Set Desktop Wallpaper
This doesn't make it functional for versions of OS X
Mozilla tracked an issue in Bug #1629313 citing OS X 10.14+. So this busted for those versions on all UXP applications until this is worked around/fixed/mangled.
More data on exact OS Versioning is need as well as a full on Mac developer.